Passing basic test for SolrQParser
Successfully performing geographically restricted searches in test. I'm
not sure why the bounding box filter works - it's using a SortableDouble
representation for the numbers, but it does seem to work. Maybe the trie
tokenizer indexes that representation as one of the tokens? Or maybe
it's just working by coincidence : \

Anyway, would like to use the NumericRangeQuery instead - I think that
is meant to work with trie fields - but first crack didn't seem to be
working.