Working distance filter with trie bounding box
Turned out the reason it was working before was that I was returning a
MatchAllDocsQuery for the starting filter. Had the wrong constant for
the ratio between degrees and miles, which is why the actual filter
wasn't working. Now it is.

Next step is to reflect via Solr on the fields being searched to
determine whether a trie or regular range query should be performed; and
if a trie query, what the precision step should be.