More final, unified implementation of distance filter/sort
Distance filtering and sort is now unified into a SpatialQueryComponent.
This component attaches a LatLongDistanceFilter to the main query (by
means of a FilteredQuery) and also, if "sort" is specified, sorts the
results using the distance filter as a starting point. The information
for the query is passed using a special "spatial" param which is in the
standard LocalParams format.