Use a filter rather than wrapping the query in a FilteredQuery
The previous approach - wrapping the main query in a FilteredQuery with
the distance filter - was causing exceptions with non-trivial query
parameters when attempting to sort. This appeared to be because the
filter was not calculating distances for all relevant docs, and the
distance sort needed that. So, instead, apply the DistanceFilter to the
Solr request's filter chain (an 'fq').